+++ example-context/example-service/example-command/example-command-api/src/main/java/uk/gov/justice/services/example/cakeshop/command/api/AddRecipeCommandApi.java
@@ -9,26 +9,17 @@ import uk.gov.justice.services.messaging.JsonEnvelope;
 
 import javax.inject.Inject;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
 /**
  *
  */
 @ServiceComponent(COMMAND_API)
 public class AddRecipeCommandApi {
 
-    private static final Logger LOGGER = LoggerFactory.getLogger(AddRecipeCommandApi.class);
-    private static final String FIELD_RECIPE_ID = "recipeId";
-
     @Inject
     Sender sender;
 
     @Handles("cakeshop.add-recipe")
     public void addRecipe(final JsonEnvelope command) {
-        LOGGER.info("=============> Inside add-recipe Command API. RecipeId: " + command.payloadAsJsonObject().getString(FIELD_RECIPE_ID));
-
         sender.send(command);
     }
-
 }