+++ example-context/example-service/example-command/example-command-api/src/main/java/uk/gov/justice/services/example/cakeshop/command/api/MakeCakeCommandApi.java
@@ -1,6 +1,5 @@
 package uk.gov.justice.services.example.cakeshop.command.api;
 
-import static org.slf4j.LoggerFactory.getLogger;
 import static uk.gov.justice.services.core.annotation.Component.COMMAND_API;
 
 import uk.gov.justice.services.core.annotation.Handles;
@@ -10,20 +9,14 @@ import uk.gov.justice.services.messaging.JsonEnvelope;
 
 import javax.inject.Inject;
 
-import org.slf4j.Logger;
-
 @ServiceComponent(COMMAND_API)
 public class MakeCakeCommandApi {
-
-    private static final Logger LOGGER = getLogger(MakeCakeCommandApi.class);
 
     @Inject
     Sender sender;
 
     @Handles("cakeshop.make-cake")
     public void handle(final JsonEnvelope envelope) {
-        LOGGER.info("=============> Inside make-cake Command API");
-
         sender.send(envelope);
     }
 }