+++ example-context/example-service/example-command/example-command-api/src/main/java/uk/gov/justice/services/example/cakeshop/command/api/OrderCakeCommandApi.java
@@ -9,21 +9,14 @@ import uk.gov.justice.services.messaging.JsonEnvelope;
 
 import javax.inject.Inject;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
 @ServiceComponent(COMMAND_API)
 public class OrderCakeCommandApi {
-
-    private static final Logger LOGGER = LoggerFactory.getLogger(OrderCakeCommandApi.class);
 
     @Inject
     Sender sender;
 
     @Handles("cakeshop.order-cake")
     public void orderCake(final JsonEnvelope command) {
-        LOGGER.info("=============> Inside order-cake Command API. ");
-
         sender.send(command);
     }
 }